move vov.nvim classobj plug back to `c` from `s`
I initially felt that `s` was more appropraite for coc.nvim lsp's
classobj since I don't work on classes as much as I do with structs etc.
I thought since I don't use `s` as a text object that often it was okay
to override, but it's not that smooth and I want to to use `s` again as
a sentence text object as it was intend and will just use `c` as
recommended by coc.nvim for classobjects.