make vim split mapping similar to tmux's
I do not use tpope/vim-vinegar enough to let it claim `-` to itself.
It does a better job being used as split window mappings so they are
similar to tmux now. Given that I don't like vim-vinegar currently, this
doesn't seem as big a hassle as I thought it woudl be.