re-appropriate nvim keymaps among oil and neotree
I use Neo-Tree way more than netrw/vim-vinegar/oil; it's just easier to
handover the previous `-`/`_` keymaps to it and losing netrw/vinegar is
not that big a deal, I seldom use them two. As for oil, I will likely
use it a lot, so assigning it to `<m-o>`/`<m-O>` seems appopriate as it
presents a gui anyway and also won't be faraway yet still.