fix(fish): remove long option names from coreutils
coreutils like `tr` cannot handle long option names in macos.
I'd prefer the long options but have to choose short.
The best option wuld be install gnu-coreutils on here, but they all
will then start with letter `g` on macos and so the commands will become
non-cross-platform. better to make the compromise and use the shorter
forms, and just add comments if necessary.